Remove types.ts and move type definitions into other files
